+++ images/controller/src/pkg/controller/nfs_storage_class_watcher.go
@@ -23,11 +23,10 @@ import (
 	"reflect"
 	"time"
 
-	"d8-controller/internal"
 	"d8-controller/pkg/config"
 	"d8-controller/pkg/logger"
 	v1alpha1 "github.com/deckhouse/csi-nfs/api/v1alpha1"
-	utilsvalidating "github.com/deckhouse/csi-nfs/lib/go/utils/pkg/validating"
+	commonvalidating "github.com/deckhouse/csi-nfs/lib/go/common/pkg/validating"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/storage/v1"
 	k8serr "k8s.io/apimachinery/pkg/api/errors"
@@ -111,7 +110,7 @@ func RunNFSStorageClassWatcherController(
 					return reconcile.Result{}, err
 				}
 
-				if err := utilsvalidating.ValidateNFSStorageClass(nfsModuleConfig, nsc, internal.FeatureTLSEnabled); err != nil {
+				if err := commonvalidating.ValidateNFSStorageClass(nfsModuleConfig, nsc); err != nil {
 					log.Error(err, "[NFSStorageClassReconciler] invalid NFSStorageClass")
 					return reconcile.Result{}, err
 				}
