+++ images/controller/src/pkg/controller/nfs_storage_class_watcher_func.go
@@ -25,9 +25,9 @@ import (
 	"strconv"
 	"strings"
 
-	"d8-controller/internal"
 	"d8-controller/pkg/logger"
 	v1alpha1 "github.com/deckhouse/csi-nfs/api/v1alpha1"
+	commonfeature "github.com/deckhouse/csi-nfs/lib/go/common/pkg/feature"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/api/storage/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -590,7 +590,7 @@ func GetSCMountOptions(nsc *v1alpha1.NFSStorageClass) []string {
 		mountOptions = append(mountOptions, "nfsvers="+nsc.Spec.Connection.NFSVersion)
 	}
 
-	if internal.FeatureTLSEnabled {
+	if commonfeature.TLSEnabled {
 		if nsc.Spec.Connection.Mtls {
 			mountOptions = append(mountOptions, "xprtsec=mtls")
 		} else if nsc.Spec.Connection.Tls {
