+++ images/webhooks/src/handlers/nscValidator.go
@@ -5,13 +5,12 @@ import (
 	"fmt"
 
 	cn "github.com/deckhouse/csi-nfs/api/v1alpha1"
-	utilsvalidating "github.com/deckhouse/csi-nfs/lib/go/utils/pkg/validating"
+	commonvalidating "github.com/deckhouse/csi-nfs/lib/go/common/pkg/validating"
 	"github.com/slok/kubewebhook/v2/pkg/model"
 	kwhvalidating "github.com/slok/kubewebhook/v2/pkg/webhook/validating"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/klog/v2"
-	"webhooks/internal"
 )
 
 const (
@@ -40,7 +39,7 @@ func NSCValidate(ctx context.Context, arReview *model.AdmissionReview, obj metav
 		klog.Fatal(err)
 	}
 
-	if err := utilsvalidating.ValidateNFSStorageClass(nfsModuleConfig, nsc, internal.FeatureTLSEnabled); err != nil {
+	if err := commonvalidating.ValidateNFSStorageClass(nfsModuleConfig, nsc); err != nil {
 		klog.Error(err)
 		return &kwhvalidating.ValidatorResult{
 			Valid:   false,
