+++ lib/go/common/pkg/feature/ee.go
@@ -13,6 +13,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package internal
+package feature
 
-const FeatureTLSEnabled = true
+const TLSEnabled = true
