--- lib/go/utils/README.md
@@ -1,6 +1,0 @@
-To use this module, execute
-```
-go get github.com/deckhouse/MODULE/api@main. 
-```
-
-The pseudo-tag will be generated automatically.
